requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
pdfminer.six
chromadb
faiss-cpu
langchain-community
langchain_google_genai